tracking-manager/packages/recommendation/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11.3
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.0
